New attempt, new problem.
I had a functioning solution for 650,
But it was taking way too long. My computer ended up shutting down.
After every turn, I need to write the answer to a file so that it
continue if it turns back on.
035 is about circular primes. I'm not sure why it's not working really,
I need to check that the primes are circulating correctly, especially as
the number gets larger.

All the best,
MJ :/