Project Euler Problems.
I made a lot of updates to the problems for project euler.
For one, I made a new attempt at 650. I liked the concept, but it took
up too much memory.
I fixed what was wrong with 35, I hadn't converted it back to an int.
As well, I added and subsequently completed 36. I worked on my
commenting skills, and used methods to complete the problem concisely.
I added 37, with the basics, yet to be completed, though it sounds like
fun.

All the best,
MJ :/