updated 650
I found that a major problem with my previous iteration was that I
couldn't close out of the window that had the computation running,
otherwise it would lose it's place.
I would have to start all over, as I did.
Now, I made a file to hold all of the sums up to n.
That way, I can iterate through that file to get to my previous position
As well, positions already computed are easily accessable now, as it
doesn't have to compute everytime.

All the best,
MJ :/