Up to 41
In this update, I created and solved 39 and 40. They were simple
problems overall, including pythagreum theorum and pandigitality.
41 was kind of intimidating at first, but I think I found a good
solution. The code I wrote could definitely be improved upon for it's
individual use, but I wanted to make a file that holds a large amount of
primes, and decided this would be a simple way to add onto my list.
41 now adds primes, and probably will until around 7654321.
If need be, this program can run again with a different target to get
larger primes. As well, the program runs faster up when inputing from
the file instead of manually making the prime list everytime.

All the best,
MJ :/