So close to 50!
A bunch of files here that I definitely have not updated on github.
AllPrimes.txt holds all of the primes I've possibly calculated. Problem 44
had to go through a lot of primes, so I documented them for possible
future use. I split this differently from Primes.txt because I wanted
650 to only use the primes it needed, not a million or more.
650 is going pretty well. On my fear trip, I let it run, and it got up
to around 18000.
45 was about triangle, pentagonal, and hexagonal numbers. This was
fairly simple, as I had completed problems with these functions before.
46 was disproving a conjecture about all odd numbers being twice a
number plus a prime. I didn't want to use the prime files I had, so I
made the similar addPrime method. I also didn't like the inefficiency of
the square root, as I know sometimes doing that will give me roundoff
errors. Thus, I had a seperate array that held perfect squares of
numbers to check after a prime was subtracted from the odd number.
47 was about the number of distinct factors of a set of sequencial
numbers. At first, I was planning on using the method of factoring that
I used in 650, but I didn't think it was necessary. Plus, I didn't want
to bring in the primes again and just make everything more complicated
than it had to be. I made the method to check if the number of factors
equaled the target, and made sure all of the numbers in the
sequence(length target) had target amount of factors. I accidentaly
submitted the wrong answer first time, as I thought we needed to submit
the sum, not the first number.
Anyways, I started of 48, but haven't really worked on it. I think I
know what to do though. I'm thinking of calculated the number to the
power of itself, then adding the last ten digits of it to a variable,
and the only taking the last ten digits of that number.

I think that's it.

All the best,
MJ :/