Solved 5ish problems
FINISHED 650!!!!!!!!!!!
Took FOREVER but I got the right answer. For future reference, look at
the page with other ppl's comments to learn more mathy things.
Lots of good methods on there, but need to make sense of them.

53 just added, no work yet, seems ok.

Most of the problems were pretty easy overall. 52 I just finished in a
few minutes, experimenting with different styles of commenting. Pretty
easy overall, worked with digits. Brute force approach.

51 was working with families of numbers. For some reason, PE actually
scored it as harder than 650. Still need to learn more about factoring.

50 was a bit problematic at first, because I was having some trouble
with a part of the problem. I had to find which prime under 1 mil can be
written as the longest series of consecutive primes. it was weird at
first, I got weird answers. Found a simpler way to save on time,
executed much faster.

49 worked with permutations and primes that increase linearly. Overall
pretty ok, used built-in sort method to sort the string to compare the
different digits. Might use this tactic in future.

I think that was it.
All the best,
MJ :/