Lots of new problems.
Sorry for the delay, if anyone even reads these or is a fanatic of my
work.
Thanks, I guess.
Anyways, I completed a lot of problems. Most were pretty easy overall.

First off, the deleted files were for taxes, so nothing to worry about
there.

52 is about a series of positiive integers that are in an arithmetic
sequence and all have the same digits. Overall, it was pretty easy. I
just had a loop and then looped inside of that to do the sequence.

53 was about combinations over 1mil. I have had experience with
combinations from problem 650, so I easily adapted to the problem. I
also remembered combinations being mirroring, so I took a shortcut to
find the answers.

54 is my magnum opus in this commit. It features a program that will
determing if a set of cards will result in player 1 winning or player 2
in a game of poker. It was really fun, and I used a lot of different
methods to get all of the different possible teirs of poker.

55 is about palindrome numbers and lychrel numbers. Again, a pretty
straight forward problem, I used a few methods to tell whether it was a
palindrome or a lychrel, and had a simple main method.

56 is a simple program that sums all of the digits of a group of
numbers, and finds which base and exponent will give you the biggest sum
of digits. I tried a new technique in this one, and was pretty surprised
to see it work great.

57 is approximating 2**.5. When I first saw the problem, I saw the
pattern that went on regarding the numberator and the denominator, and
could easily determine what the next num and den would be.

58 I just finished, and includes a method to tell if a number is prime
or not. I had to fidget a bit with the starting values, as I mixed one
or two up. But in the end I got the right answer.

I just started 59, and it looks kinda hard, so I'm not sure how I'll
tackle it.

I really should do more commits. Dang.
I'll try and work on this.

All the best,
MJ :/