Solved 59!
From now on, I want to commit more often, pushing when I have internet
access and when I have time.
Anyways, I got the right answer for 59. I used a variable that held the
entire alphabet and then used each letter to find the string w/ length 3
that would turn the coded message into a string. This is probably really
inefficient, as it is a triple nested for loop. thus, 26**3 loops.
The one thing that I learned from this is the use of bitwise functions,
including ^ xor, & and, | or, >> right shift, <<left shift.
pretty cool and useful stuff.
And I now can decode messages. Very cool.

All the best,
MJ :/