Unchecked Solution to 62 and 63
I think I found the answer for 62 and 63, but I can't
check it because internet is being big dumb in the basement.
For 62, I continuously generated cubes until I got the
answer. I would take each cube, turn it into a string,
and then sort the string. If the sorted string was
already a key in the dictionary, it is a permutation of a
smaller cube, so I added 1 to the number of permutations for
that exact sequence. Otherwise, if it wasn't a key, then it
is the smallest cube of that permuation, so I recorded it.
pretty small program, but only 5% difficulty, so makes
sense.
63 was nearly as simple, if not more so.
I realized that the base couldn't be larger than 10,
since base >= 10 would always have digits > power. Because
of this, for every power, I looped from 9-1 and, if the
result was less digits than the power, then I would break
and move onto the next power and reset to 9.
Sorry for no problem description on 63 yet, still need that
sweet sweet wifi to get booted better.

All the best,
MJ