Completed 64
Started and wrapped up 64 rather quickly.
I made a function to get the new a value based
off the n val, constant, and denominator.
Then, I used that to make a function to get the
entire repeated sequence.
Upon testing, it gave me the sequences unless
the sequence only had 1 value in it, so I had to
account for a null sequence as an odd one.
Otherwise, just had to work out what impacted
what.

All the best,
MJ