Completed 66 Mitchell Johnstone
Quarantine baby! Oh yeah. Day 9 of being stuck in my
room, and honestly it's getting better. If only I
actually had wifi, since 67 involves a downloaded txt
file.
anyways, for this problem, I had some trouble to start
out with, but once I found the correct approach it was
fairly straightforward.
I originally started tackling this problem with a brute
force to solve for the smallest x for each D by
incrementing x and checking if it would work. While x
increased fast, and i got rid of about 3/4 of the
possible D values, it eventually platued and I couldn't
make it any more efficient. After trying similar ideas,
but using y and D for my incremented variables, I needed
another way to do this. After looking up what the
equation was, it turns out that minimal x was related
to the square root of D using the convergences of the
infinite continued fractions of the square root of D.
Since 64 and 65 involved generating the pattern for the
continued fraction and solving the convergence for the
continued fraction, respectively, I used ideas from both
to get a program that worked efficiently at the end.
One hell of a beast, but tamed once I learned.

All the best,
MJ