Solved P67. Yeet Mitchell Johnstone
Alright, back at it again with another
project Euler problem. Yet again, it was
a good warm up from skills I've been
behind on, such as working with files.
However, after a quick refresher, I
remembered how to do the simpler version
of this problem, by approaching it
backwards. After setting it up, it was
quite simple. Ready for the next one!

All the best,
MJ