Solved P68. Mitchell Johnstone
Arise! WiFi! it has been aquired! HUZZAH!!
Well, as the title suggests, I also solved 68, tho it's
less exciting than getting that sweet sweet wifi.
First, I found all permutations of 1-10. Then,
when I looked at the base case, the 6 nodes with 1-6, I
found that the resulting nodes followed a pattern in
its construction, so I used a similar pattern for the
15 nodes.
After that, I used the conditions that the problem set,
that the first node is the smallest external node, and
that all triplets add to the same total, to quickly
get rid of many of the permutations. I let it run, and
it took about 7 seconds for the final answer.

All the best,
MJ