Solved P69. Mitchell Johnstone
Yuh-Yeet.
What a day.
I'm so tired, yet still got this working.
Had to make a phi function, but found a way to change the
traditional sieve to give me the lowest prime factor of a
number. This let me get all the factors of a number pretty
quickly, which I used in an equation I found from research.
BEANS.

All the best,
MJ