Solved P70 Mitchell Johnstone
Fairly straight forward.
I wanted to compress it to one single line, but couldn't
figure out how to output a different number other than
what I was iterating over. However, since I had used
the totient function in P069, It was easy to simply use
that method in this problem, and do a simple search for
a min.

All the best,
MJ