Vim and P071 MJ
This is the second time I've tried using Vim, but this time much
more dedicated. It's actually really fun to use once I got the hang
of it, but I still need to get better. Anyways, only a small start
on 71. I'm thinking of using a class to better get the decimal
approx of each of the numbers.
For now though, I hope you have a great day.

All the best,
MJ :/