Solved problem 71
I overcomplicated this whole problem upon review.
I looked at the thread about this problem, and nearly everyone
used just paper and pencil. To contrast, I used a brute force
with a binary search. However, I did learn about a Farey series,
which is cool and probably useful for the next few problems.
Otherwise, I had some debugging to do on this for a while
that was frustrating, but got it working. I used Vim to completely
write this file, which was really really cool and fun to do and
use. I will try to use this more :)

All the best,
MJ :/