Solve 72
Today I started and finished problem 72.
At first, being so similar to 71, I started
trying to improve the process of 71.
It was a good endeavor, but probably better
for problem 73. In a eureka moment, I realized
that for each denominator, the reduced proper fractions have
numerators that were relatively prime to the denominators.
revisiting problem 70, I then used the totient function
to find the number of relatively prime numerators and summed them
all in a quick 1 line function. The first method would've worked,
but would've taken many minutes, while the one-line took only
a few seconds.
All in all, a lot of fun :)

All the best,
MJ :/