Solve 73, fix method in 72
After doing prior prep work in the beginning of problem 72, 73
was simple enough, just reusing the methods. However, after
working the math, I had to fix a small error in the minimum
numerator method from 72. After the fix, it was only a couple of
nested loops resulted in a couple seconds and the right answer.

All the best,
MJ :/