Solve 74
Overall, a brute force method was what I opted for.
I thought of looking for the chains as the terminal points,
but that would only take off a few milliseconds each. Seemed like
a waste. But overall it was straightforward after reading through
it to make sure I understood it. A nice one-line method and a good
loop, and all was solved.

All the best,
MJ :/