Finish 75, customized vim
This one took WEEKS.
Sorry for that :/
but hey, I got it finished!
I had to write like 6 different methods for the problem, but they were
all taking so long. Finally, I let one run for long enough, and it gave
me the right answer after a lot of debugging and wrong answers.
Then, when reviewing the code and other people's suggestions, it dawned
on me that I was being dumb with searching in the arrays. While
theoretically it would be faster to store it the way I did in v6, it's
much faster to just access indices instead of doing a full search,
because by searching you essentially have to access half the values on
average. Which adds up over time. Ugh.
Anyways, I added a 7th version afterwards that addressed this problem,
using the recursive approach in 6, but now storing the perimeters much
more efficiently, and it took merely 2 seconds.

Otherwise, there are hundreds of other files added in this commit, and
that is because I was customizing vim. It looks dope now, but I needed
to do some work in a virtual environment to get the autocomplete
working, which it is :)
sorry for all the files tho.

All the best,
MJ :/