Solved 76
I initially wanted to make a recursive program, since that what it
looked like in the test case. However, while setting up the recursive
function, I made a function to generate the next partitioning, and
realized the recursive part could be swapped for an iterative approach.
After solving it with true counting, I read through the thread for the
problem and found out about "partitioning equations", one which involved
finding the number of ways to split a number m into integers. I used
that thread, along with recursion, to generate the improvement that took
milliseconds to complete.
The next problem is similar but with primes, so I will be trying a
similar approach to start with but include some tracking of solutions
for prior numbers... gonna be fun

All the best,
MJ :/