cleaning up repo, generating problem program
There were some unnecessary files created by vim, which I deleted. I
also made a new program, GenerateProblems!! I'm excited about it because
I wrote it in roughly 2 hours and it works great. Essentially, it crawls
the project euler website, takes the problems I don't already have
solutions for, and makes a new file for them. That way, I already have
the info for the problem in the file, as well as a start to the basic
template, so I don't have to rewrite stuff everytime. P001 and P002 are
my test files, and they worked fantastically.
Well, gonna go update all my files for all of the problems!
See you in the next commit.

All the best,
MJ :/