Added missing problems, fix generating length
I fixed the generating code so it allows for 100 characters in a line.
It was messing with problem 8 to only have 50 chars, so I figured 100
would be reasonable. Otherwise, I ran the program with success, as
denoted by the many new problems I have generated files for.

All the best,
MJ :/