Finished 77 and 78
in reality, 76, 77, and 78 were all very similar problems. They all
involved partitioning integers, just in different ways. Actually, 78 was
the same as 76, just it included itself as a partition. For 77, I looked
up different ways for partitioning a number into primes, and found a
simple elegant solution. I used the prime list text file I used so I
didn't have to generate them myself.
As for 78, I spent a lot more time researching different ways to
calculate the number of partitions. I started with the improvement from
76, but then delved into some other functions from online, including
recursive formats and generating the actual partitions. Since the 2nd
version was the fastest, i let it run overnight, but something must be
wrong with it because it skipped the correct answer. I created another
program after that, following a different method, and it was effective.
It stopped much earlier than the original program, so I let it run again
to double check. Then, because it was all addition, I added mods to
speed it a bit, and came up with the right answer.

All the best,
MJ :/